+++ Sources/CommandRegistry/TypeAliases.swift
@@ -1,0 +1,11 @@
+//
+//  TypeAliases.swift
+//  CommandRegistry
+//
+//  Created by Eneko Alonso on 3/9/19.
+//
+
+import Foundation
+import Utility
+
+public typealias ArgumentParser = Utility.ArgumentParser
